update: fix precision
